refactor: move file to util
룸매니저, 룸, 유저 클래스 util로 이동
